requirements.txt
aiofile==3.9.0
aiofiles==24.1.0
alembic==1.9.3
altair==6.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==26.1.0
audioop-lts==0.2.2
Authlib==1.6.9
bcrypt==5.0.0
beartype==0.22.9
billiard==4.2.4
blinker==1.9.0
brotli==1.2.0
cachetools==7.0.5
caio==0.9.25
celery==5.3.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.6
cyclopts==4.10.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
ecdsa==0.19.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.3
fastapi-cli==0.0.24
fastmcp==3.2.0
ffmpy==1.0.0
filelock==3.25.2
fsspec==2026.3.0
gitdb==4.0.12
GitPython==3.1.46
gradio==6.11.0
gradio_client==2.4.0
groovy==0.1.2
h11==0.16.0
hf-gradio==0.3.0
hf-xet==1.4.3
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.9.0
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.27.0
mdurl==0.1.2
more-itertools==11.0.1
narwhals==2.18.0
notion-client==3.0.0
numpy==1.26.4
openai==2.30.0
openapi-pydantic==0.5.1
openenv-core==0.2.3
opentelemetry-api==1.40.0
orjson==3.11.8
packaging==26.0
pandas==2.2.1
passlib==1.7.4
pathable==0.5.0
pillow==12.1.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
protobuf==6.33.6
psycopg2-binary==2.9.10
py-key-value-aio==0.4.4
pyarrow==23.0.1
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
Pygments==2.20.0
PyJWT==2.12.1
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.3.0
python-multipart==0.0.22
python-telegram-bot==22.7
pytz==2026.1.post1
PyYAML==6.0.3
redis==4.5.4
referencing==0.37.0
requests==2.32.5
rfc3986==1.5.0
rich==14.3.3
rich-rst==1.3.2
rich-toolkit==0.19.7
rpds-py==0.30.0
rsa==4.9.1
safehttpx==0.1.7
schedule==1.2.2
scikit-learn==1.4.1.post1
scipy==1.17.1
semantic-version==2.10.0
sentry-sdk==1.40.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==1.4.46
sse-starlette==3.3.4
starlette==1.0.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
tomli==2.4.1
tomli_w==1.2.0
tomlkit==0.13.3
tornado==6.5.5
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.12.0
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.43.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
xgboost==2.0.3
zipp==3.23.0
